Explain multiple user names
Thanks for this very helpful notebook. Minor addition, since it took me a couple tries to figure out how to add multiple fields because I couldn't find documentation on it. I cannot guarantee it's right, though, because I wasn't able to test it aside from with my own account. If you wish to be sure, I have added https://asg017.static.observableusercontent.com to the whitelist for https://bmschmidt-cors-observable.herokuapp.com for a little while so that you can test.